Rename: 테스트 파일 각각의 폴더 하위로 이동
